move out the test, avoid there components usage
